trying to group teams into leagues and divisions.
