+++ pages/scratchpad/team-listing.js
@@ -1,0 +1,6 @@
+const AmericanEast = ["TBD", "TOR", "NYY", "BAL", "BOS"];
+const AmericanCentral = ["MIN", "CLE", "CHW", "DET", "KCR"];
+const AmericanWest = ["TEX", "ANA", "SEA", "HOU", "OAK"];
+const NationalEast = ["ATL", "NYM", "FLA", "PHI", "WSN"];
+const NationalCentral = ["MIL", "PIT", "CHC", "STL", "CIN"];
+const NationalWest = ["ARI", "LAD", "SDP", "SFG", "COL"];
